refactor(game): replace dynamic imports with direct GameScene usage
Remove dynamic loading wrappers for GameScene in apps/garden and apps/www,
switching to direct imports from the game package. This simplifies the code
and improves maintainability by consolidating the dynamic import logic inside
the game package itself. Also update exports and imports accordingly to use
the new GameSceneDynamic component internally while exposing GameScene.
