+++ apps/garden/app/page.tsx
@@ -1,11 +1,11 @@
 import { SignedOut } from "@signalco/auth-client/components";
 import LoginModal from "../components/auth/LoginModal";
-import { GameSceneDynamic } from "./GameSceneDynamic";
 import { enableDebugHudFlag, enableRaisedBedDiaryFlag, enableRaisedBedFieldDiaryFlag, enableRaisedBedFieldOperationsFlag, enableRaisedBedFieldWateringFlag, enableRaisedBedOperationsFlag, enableRaisedBedWateringFlag } from "./flags";
 import { ComponentProps } from "react";
+import { GameScene } from "@gredice/game";
 
 export default async function Home() {
-  const flags: ComponentProps<typeof GameSceneDynamic>["flags"] = {
+  const flags: ComponentProps<typeof GameScene>["flags"] = {
     enableDebugHudFlag: await enableDebugHudFlag(),
     enableDebugCloseupFlag: await enableDebugHudFlag(),
     enableRaisedBedWateringFlag: await enableRaisedBedWateringFlag(),
@@ -18,7 +18,7 @@ export default async function Home() {
 
   return (
     <div className="grid grid-cols-1 h-[100dvh] relative overflow-hidden">
-      <GameSceneDynamic flags={flags} />
+      <GameScene flags={flags} />
       <SignedOut>
         <LoginModal />
       </SignedOut>
