+++ packages/game/src/indicators/GardenLoadingIndicator.module.scss
