+++ packages/game/src/indicators/GardenLoadingIndicator.tsx
